+++ lib/dradis/plugins.rb
@@ -57,6 +57,15 @@ module Dradis
         @@extensions.include?(const)
       end
 
+      def upload_integration_names_and_modules
+        with_feature(:upload).each_with_object({}) do |integration, integrations_hash|
+          integration_name = integration.plugin_name.to_s
+          integration_module = integration.module_parent
+
+          integrations_hash[integration_name] = integration_module
+        end
+      end
+
       private
 
       # Use this to ensure the Extension conforms with some expected interface
@@ -74,11 +83,12 @@ require 'dradis/plugins/engine'
 require 'dradis/plugins/version'
 
 require 'dradis/plugins/content_service/base'
-require 'dradis/plugins/template_service'
+require 'dradis/plugins/mapping_service'
 
 require 'dradis/plugins/base'
 require 'dradis/plugins/export'
 require 'dradis/plugins/import'
+require 'dradis/plugins/mappings'
 require 'dradis/plugins/upload'
 
 # Common functionality
@@ -86,6 +96,7 @@ require 'dradis/plugins/configurable'
 require 'dradis/plugins/settings'
 require 'dradis/plugins/settings/adapters/db'
 require 'dradis/plugins/settings/adapters/encrypted_configuration'
-require 'dradis/plugins/templates'
+require 'dradis/plugins/templates/samples'
+require 'dradis/plugins/templates/migrate_templates'
 require 'dradis/plugins/thor'
 require 'dradis/plugins/thor_helper'