+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v4.12.0 (XXXX 2024)
   - Update Dradis links in README
   - Fix the TypeError around the plugins template caching
+  - Remove template_service and add mapping_service to apply mappings on tool upload
+  - Include mapping module when integrations provide 'upload'
 
 v4.11.0 (January 2024)
   - No changes